backend/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
colorama==0.4.4
coverage==6.4.3
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
